+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## [0.5.72] - 2020-01-04
+### Added
+- The `admin full width` Configuration directive.
+- The `wrap lines in playground` Configuration directive.
+### Fixed
+- Bug in `legal.py`.
+- Not safe to concatenate a `DAList` and a `DAEmpty`.
+
+## [0.5.71] - 2020-01-04
+### Changed
+- Discontinued support for Python 2.7.
+
 ## [0.5.70] - 2019-12-30
 ### Added
 - The `phone_number_formatted()` function.